+++ src/DependencyInjection/HeimrichHannotEntityImportExtension.php
@@ -1,0 +1,30 @@
+<?php
+
+/*
+ * Copyright (c) 2020 Heimrich & Hannot GmbH
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+namespace HeimrichHannot\EntityImportBundle\DependencyInjection;
+
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Extension\Extension;
+
+class HeimrichHannotEntityImportExtension extends Extension
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function load(array $configs, ContainerBuilder $container)
+    {
+        $configuration = new Configuration();
+
+        $container->setParameter(Configuration::ROOT_ID, $this->processConfiguration($configuration, $configs));
+    }
+
+    public function getAlias()
+    {
+        return Configuration::ROOT_ID;
+    }
+}