+++ src/Importer/ImporterInterface.php
@@ -10,6 +10,9 @@ namespace HeimrichHannot\EntityImportBundle\Importer;
 
 interface ImporterInterface
 {
+    const LOCK_KEY = 'contao_entity_import_bundle.%s.id%s';
+    const LOCK_DIRECTORY = '/var';
+
     public function getDataFromSource(): array;
 
     public function run(): bool;